add e2k support

+++ src/node/mod.rs
@@ -35,7 +35,7 @@ impl DrmNode {
     /// Creates a DRM node from a file stat.
     pub fn from_stat(stat: Stat) -> Result<DrmNode, CreateDrmNodeError> {
         let dev = stat.st_rdev;
-        DrmNode::from_dev_id(dev)
+        DrmNode::from_dev_id(dev as u64)
     }
 
     /// Creates a DRM node from a [`dev_t`].
@@ -260,7 +260,7 @@ pub fn is_device_drm(dev: dev_t) -> bool {
 /// Returns the path of a specific type of node from the same DRM device as another path of the same node.
 pub fn path_to_type<P: AsRef<Path>>(path: P, ty: NodeType) -> io::Result<PathBuf> {
     let stat = stat(path.as_ref()).map_err(Into::<io::Error>::into)?;
-    dev_path(stat.st_rdev, ty)
+    dev_path(stat.st_rdev as u64, ty)
 }
 
 /// Returns the path of a specific type of node from the same DRM device as an existing [`DrmNode`].
